feat: Add configure Azure settings command
Add a new command to configure Azure settings for the Speechify extension. This command allows users to enter their Azure Speech Services Key and Region, which will be used for text-to-speech conversion. The new command is registered in the extension activation function and added to the package.json file for easy access in the VS Code UI.